Rename gallery to tsx
